+++ src/apps/replication/lib/mutation_log.cpp
@@ -633,7 +633,7 @@ std::map<int, log_file_ptr>& mutation_log::get_logfiles_for_test()
 /*static*/ log_file_ptr log_file::create_write(const char* dir, int index, int64_t startOffset, int maxStalenessForCommit, int writeTaskNumber)
 {
     char path[512]; 
-    sprintf (path, "%s/log.%u.%llu", dir, index, startOffset);
+    sprintf (path, "%s/log.%u.%lld", dir, index, static_cast<long long int>(startOffset));
     
     handle_t hFile = dsn::service::file::open(path, O_RDWR | O_CREAT, 0);
     if (hFile == 0)
